% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-git.R
\name{add_large_files2gitignore}
\alias{add_large_files2gitignore}
\title{Add Large Files to .gitignore}
\usage{
add_large_files2gitignore(path, size_limit = 200)
}
\arguments{
\item{path}{Character string specifying the path to the folder to be scanned.}

\item{size_limit}{Numeric value indicating the size limit in megabytes (MB).
Files larger than this limit will be added to `.gitignore`. Default is 200 MB.}
}
\value{
Prints a message indicating the number of large files added to `.gitignore`,
        or a message stating that no large files were found.
}
\description{
This function scans a specified folder for files larger than a given size limit
and appends their paths to the `.gitignore` file to prevent them from being
committed to a Git repository.
}
\examples{
\dontrun{
add_large_files2gitignore("path/to/your/folder", size_limit = 50)
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
