% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{craw_or_load_publication}
\alias{craw_or_load_publication}
\title{craw_or_load_publication}
\usage{
craw_or_load_publication(
  user_id = "3TK9yz8AAAAJ",
  pub_id = "2osOgNQ5qMEC",
  force = FALSE,
  interval = 7
)
}
\arguments{
\item{user_id}{Google Scholar user ID.}

\item{pub_id}{Publication ID.}

\item{force}{Force a fresh webpage request?}

\item{interval}{Cache validity interval in days when local data already exist.}
}
\value{
Parsed publication page data.
}
\description{
Crawl or load webpage data for a single publication.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
