% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{draw_citation_history}
\alias{draw_citation_history}
\title{draw_citation_history}
\usage{
draw_citation_history(user_id = "3TK9yz8AAAAJ", force = FALSE, interval = 7)
}
\arguments{
\item{user_id}{Google Scholar user ID.}

\item{force}{Force a fresh webpage request?}

\item{interval}{Cache validity interval in days when local data already exist.}
}
\value{
A citation history plot.
}
\description{
Draw citation history plot.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
