% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-wowchemy.R
\name{generate_collaborator_map}
\alias{generate_collaborator_map}
\title{generate_collaborator_map}
\usage{
generate_collaborator_map(
  gs_user_id = "3TK9yz8AAAAJ",
  force = FALSE,
  interval = 7,
  map_color = "#76b5c5",
  point_color = "#873e23"
)
}
\arguments{
\item{gs_user_id}{Google Scholar user ID.}

\item{force}{Force a fresh webpage request?}

\item{interval}{Cache validity interval in days when local data already exist.}

\item{map_color}{Map fill color.}

\item{point_color}{Point color.}
}
\value{
A `ggplot2` object.
}
\description{
Generate a collaborator map.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
