% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-openwd.R
\name{openwd}
\alias{openwd}
\title{openwd}
\usage{
openwd()
}
\value{
inder.
}
\description{
Open current work directory.
}
\examples{
\dontrun{
openwd()
}
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
