% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{request_publication_info}
\alias{request_publication_info}
\title{request_publication_info}
\usage{
request_publication_info(
  user_id = "3TK9yz8AAAAJ",
  pub_id = "0EnyYjriUFMC",
  force = FALSE,
  interval = 7
)
}
\arguments{
\item{user_id}{Google Scholar user ID.}

\item{pub_id}{Google Scholar publication ID.}

\item{force}{Force a fresh webpage request?}

\item{interval}{Cache validity interval in days when local data already exist.}
}
\value{
Publication information.
}
\description{
Request information for a single publication.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
