% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-github-gitlab-gitee.R
\name{require_github_profile}
\alias{require_github_profile}
\title{require_github_profile}
\usage{
require_github_profile(user_name = "jaspershen")
}
\arguments{
\item{user_name}{GitHub user name.}
}
\value{
A list containing profile information.
}
\description{
Request a GitHub profile.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
