% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-utils.R
\name{show_progresser}
\alias{show_progresser}
\title{show_progresser}
\usage{
show_progresser(
  index = 1:1000,
  progresser = c(1, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100)
)
}
\arguments{
\item{index}{A numeric index vector.}

\item{progresser}{Progress labels to map onto `index`.}
}
\value{
A data frame.
}
\description{
Create a simple progress lookup table.
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
