% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_ggplot2_themes.R
\name{theme_shenlab}
\alias{theme_shenlab}
\title{Shen Lab Custom ggplot2 Theme}
\usage{
theme_shenlab(base_size = 12)
}
\arguments{
\item{base_size}{Numeric. Base font size for text elements (default = 12).}
}
\value{
A \code{\link[ggplot2]{theme}} object that can be added to ggplot2 plots.
}
\description{
A simple custom ggplot2 theme based on \code{\link[ggplot2]{theme_bw}}
with minor grid lines removed and axis text sizes adjusted.
}
\details{
This theme builds on \code{\link[ggplot2]{theme_bw}} and customizes:
\itemize{
  \item Removes minor grid lines.
  \item Sets axis titles and axis text to \code{base_size}.
}
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + theme_shenlab()

}
