% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-wowchemy.R
\name{update_packages_tidymass_org}
\alias{update_packages_tidymass_org}
\title{update_packages_tidymass_org}
\usage{
update_packages_tidymass_org(
  path_tidymass.org = "tidymass-website/static/tidymass-packages",
  path_packages = "."
)
}
\arguments{
\item{path_tidymass.org}{Path to the `tidymass-packages` directory.}

\item{path_packages}{Path containing package source tarballs.}
}
\value{
Invisibly updates package files in `path_tidymass.org`.
}
\description{
Update package files used by tidymass.org.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
