% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-pubmed.R
\name{convert_gs_pub_id2pmid}
\alias{convert_gs_pub_id2pmid}
\title{convert_gs_pub_id2pmid}
\usage{
convert_gs_pub_id2pmid(
  gs_user_id = "3TK9yz8AAAAJ",
  gs_pub_id = "0EnyYjriUFMC",
  force = FALSE,
  interval = 7
)
}
\arguments{
\item{gs_user_id}{google scholar user ID}

\item{gs_pub_id}{google scholar publication ID}

\item{force}{Force read webpage?}

\item{interval}{If there is old data on the local machine,
do you want to use it and the interval (day)?}
}
\value{
Information of publication
}
\description{
Convert google scholar publication ID to pubmed ID
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
