% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{craw_or_load_page}
\alias{craw_or_load_page}
\title{craw_or_load_page}
\usage{
craw_or_load_page(user_id = "3TK9yz8AAAAJ", force = FALSE, interval = 7)
}
\arguments{
\item{user_id}{Google Scholar user ID.}

\item{force}{Force a fresh webpage request?}

\item{interval}{Cache validity interval in days when local data already exist.}
}
\value{
Parsed page data.
}
\description{
Crawl or load webpage data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
