% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{craw_or_load_page}
\alias{craw_or_load_page}
\title{craw_or_load_page}
\usage{
craw_or_load_page(user_id = "3TK9yz8AAAAJ", force = FALSE, interval = 7)
}
\arguments{
\item{user_id}{User google scholar ID}

\item{force}{Force read webpage?}

\item{interval}{If there is old data on the local machine,
do you want to use it and the interval (day)?}
}
\value{
page data
}
\description{
Craw or load webpage data.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
