% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{craw_or_load_publication}
\alias{craw_or_load_publication}
\title{craw_or_load_publication}
\usage{
craw_or_load_publication(
  user_id = "3TK9yz8AAAAJ",
  pub_id = "2osOgNQ5qMEC",
  force = FALSE,
  interval = 7
)
}
\arguments{
\item{user_id}{User google scholar ID}

\item{pub_id}{publication ID}

\item{force}{Force read webpage?}

\item{interval}{If there is old data on the local machine,
do you want to use it and the interval (day)?}
}
\value{
page data
}
\description{
Craw or load webpage data of one publication.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
