% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-system_tools.R
\name{create_project_organization}
\alias{create_project_organization}
\title{create_project_organization}
\usage{
create_project_organization()
}
\value{
Creates a standard set of project folders and template files.
}
\description{
Create a standard project directory structure.
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
