% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-system_tools.R
\name{create_project_organization}
\alias{create_project_organization}
\title{create_project_organization}
\usage{
create_project_organization()
}
\value{
some folders
}
\description{
Create project organization.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
