% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{draw_citation_history}
\alias{draw_citation_history}
\title{draw_citation_history}
\usage{
draw_citation_history(user_id = "3TK9yz8AAAAJ")
}
\arguments{
\item{user_id}{User google scholar ID}
}
\value{
Profile.
}
\description{
Draw citation history plot.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
