% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-system_tools.R
\name{get_os}
\alias{get_os}
\title{get_os}
\usage{
get_os()
}
\value{
OS name.
}
\description{
Get operating system information.
}
\examples{
get_os()
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
