% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-get_project_wd.R
\name{get_project_wd}
\alias{get_project_wd}
\title{get_project_wd}
\usage{
get_project_wd()
}
\value{
A project working directory path.
}
\description{
Get the project working directory containing the `.Rproj` file.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
