% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-utils.R, R/9-git.R
\name{install_fastgit}
\alias{install_fastgit}
\title{install_fastgit}
\usage{
install_fastgit(pkg, from = c("gitee", "gitlab", "github"), ...)

install_fastgit(pkg, from = c("gitee", "gitlab", "github"), ...)
}
\arguments{
\item{pkg}{Package name from GitHub, GitLab, or Gitee in `"name/repo"` format.}

\item{from}{One of `"gitee"`, `"gitlab"`, or `"github"`.}

\item{...}{Other parameters passed to `remotes::install_git()`.}
}
\description{
Install packages from FastGit. Credit to Shixiang Wang.

Install packages from FastGit. Credit to Shixiang Wang.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
