% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-r4projects_logo.R
\name{r4projects_logo}
\alias{r4projects_logo}
\title{r4projects_logo}
\usage{
r4projects_logo()
}
\value{
r4projects logo
}
\description{
Get the detailed of metPath package.
}
\examples{
r4projects_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
