% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{request_citation}
\alias{request_citation}
\title{request_citation}
\usage{
request_citation(user_id = "3TK9yz8AAAAJ", force = FALSE, interval = 7)
}
\arguments{
\item{user_id}{Google Scholar user ID.}

\item{force}{Force a fresh webpage request?}

\item{interval}{Cache validity interval in days when local data already exist.}
}
\value{
Citation summary data.
}
\description{
Request citation information.
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
