% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{request_profile}
\alias{request_profile}
\title{request_profile}
\usage{
request_profile(user_id = "3TK9yz8AAAAJ", force = FALSE, interval = 7)
}
\arguments{
\item{user_id}{User google scholar ID}

\item{force}{Force read webpage?}

\item{interval}{If there is old data on the local machine,
do you want to use it and the interval (day)?}
}
\value{
Profile.
}
\description{
Request the profile.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
