% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{request_profile}
\alias{request_profile}
\title{request_profile}
\usage{
request_profile(user_id = "3TK9yz8AAAAJ")
}
\arguments{
\item{user_id}{User google scholar ID}
}
\value{
Profile.
}
\description{
Request the profile.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
