% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{request_publication_info}
\alias{request_publication_info}
\title{request_publication_info}
\usage{
request_publication_info(
  user_id = "3TK9yz8AAAAJ",
  pub_id = "0EnyYjriUFMC",
  force = FALSE,
  interval = 7
)
}
\arguments{
\item{user_id}{User google scholar ID}

\item{pub_id}{Publication google scholar ID}

\item{force}{Force read webpage?}

\item{interval}{If there is old data on the local machine,
do you want to use it and the interval (day)?}
}
\value{
Information of publications
}
\description{
Request one publication info
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
