% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-google_scholar.R
\name{request_publications}
\alias{request_publications}
\title{request_publications}
\usage{
request_publications(user_id = "3TK9yz8AAAAJ")
}
\arguments{
\item{user_id}{User google scholar ID}
}
\value{
Information of publications
}
\description{
Request all publications
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
