% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-pubmed.R
\name{request_pubmed_publication_info}
\alias{request_pubmed_publication_info}
\title{request_pubmed_publication_info}
\usage{
request_pubmed_publication_info(pmid = "35902589")
}
\arguments{
\item{pmid}{pmid of paper}
}
\value{
Information of publication
}
\description{
Request publication information from pubmed
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
