% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-github-gitlab-gitee.R
\name{require_github_contribution_activity}
\alias{require_github_contribution_activity}
\title{require_github_contribution_activity}
\usage{
require_github_contribution_activity(user_name = "jaspershen")
}
\arguments{
\item{user_name}{GitHub user name}
}
\value{
Profile.
}
\description{
Request the contribution activity of github
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
