% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-github-gitlab-gitee.R
\name{require_package_info}
\alias{require_package_info}
\title{require_package_info}
\usage{
require_package_info(
  repo_name = "jaspershen/laggedcor",
  from = c("github", "gitlab", "gitee")
)
}
\arguments{
\item{repo_name}{R package repository name. Defaults to `"jaspershen/laggedcor"`.}

\item{from}{One of `"github"`, `"gitlab"`, or `"gitee"`.}
}
\value{
information
}
\description{
Request R package information from GitHub, GitLab or Gitee.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
