% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-wowchemy.R
\name{update_packages_tidymass_org}
\alias{update_packages_tidymass_org}
\title{update_packages_tidymass_org}
\usage{
update_packages_tidymass_org(
  path_tidymass.org = "tidymass-website/static/tidymass-packages",
  path_packages = "."
)
}
\arguments{
\item{path_tidymass.org}{path_tidymass.org}

\item{path_packages}{path_packages}
}
\value{
Profile.
}
\description{
Update packages in tidymass.org. Please set the path to "/Users/xiaotaoshen/tidymass" first.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
